fix: 修正Histogram的Observe方法参数类型并更新运行目录
将SafeHistogram的Observe方法参数从float64改为int64以匹配实际使用需求
更新run.ps1脚本中的metrics目录路径

+++ internal/template/template.go
@@ -159,7 +159,7 @@ type SafeHistogram{{ if .Labels }}_{{ .Name | toPascal }}{{ end }} struct {
 
 {{- if hasMethod .Methods "observe" }}
 // Observe records the observed value.
-func (h *SafeHistogram{{ if .Labels }}_{{ .Name | toPascal }}{{ end }}) Observe(v float64, {{ labelParams .Labels }}) {
+func (h *SafeHistogram{{ if .Labels }}_{{ .Name | toPascal }}{{ end }}) Observe(v int64, {{ labelParams .Labels }}) {
     h.HistogramVec.Observe(v, {{ labelArgs .Labels }})
 }
 {{- end }}
